Reddit

Documentation

data Credentials

Constructors

Credentials

Fields

username :: Text

password :: Text

clientID :: Text

clientSecret :: Text

data TokenInternal

Constructors

TokenInternal

Fields

_access_token :: Text

_token_type :: Text

_expires_in :: Int

_scope :: Text

data Token

Constructors

Token

Fields

token :: ByteString

token_type :: ByteString

expires_at :: UTCTime

scope :: ByteString

convertToken :: TokenInternal -> IO Token

data Env

Everything required to query the Reddit API.

Constructors

Env

Fields

envToken :: ByteString

envTokenType :: ByteString

envTokenExpiresAt :: UTCTime

envTokenScope :: ByteString

envUserAgent :: ByteString

authWithCredentials :: Credentials -> Text -> IO Env

Get an Env value by supplying user account credentials. Only to be used for quick scripts, etc.

oauth :: Text

withUAToken :: Env -> Option 'Https

data RedditException

Constructors

TokenExpiredException

checkTokenValidity :: ReaderT Env IO ()

TODO: Don't check before using it, check the response headers instead to determine if token has expired

user :: Text -> ReaderT Env IO ByteString

Get information about a user

data Timeframe

Constructors

Hour

Day

Week

Month

Year

All

data SubredditSort

Constructors

Hot

New

Random

Rising

Top Timeframe

Controversial Timeframe

subreddit :: Text -> SubredditSort -> ReaderT Env IO ByteString

Get the first 25 posts on a subreddit

subComments :: Text -> ReaderT Env IO ByteString

Get the most recent 25 comments on a subreddit. This endpoint is undocumented!