Reddit.Auth

Credentials

This section describes the credentials needed for using the Reddit API as a 'script'.

In order to obtain these, you will need to:

Log in using the account you want to post or query Reddit as (this may be a bot account, for example)

Navigate to https://www.reddit.com/prefs/apps

Click 'create another app' at the bottom and select the 'script' radio button

Fill in the required details. For a personal script, the redirect URI is not important, set it to anything you like.

Take note of the app ID (a string with random characters) and the secret (the same but longer). These become, respectively, clientID and clientSecret.

username and password are the Reddit login credentials of the account you are using.

Note: Do NOT ever share your password or the client secret publicly!

See also: https://github.com/reddit-archive/reddit/wiki/OAuth2-Quick-Start-Example

data Credentials

Constructors

Credentials

Fields

username :: Text

password :: Text

clientID :: Text

clientSecret :: Text

withCredentials

Arguments

:: Credentials

-> Text
Your user-agent. Reddit says you should use a unique and identifiable user-agent.

-> IO RedditEnv

Exchange user account credentials for a RedditEnv, which is required to run all Reddit queries.

data RedditEnv

Everything required to query the Reddit API.

Constructors

RedditEnv

Fields

envToken :: ByteString

envTokenType :: ByteString

envTokenExpiresAt :: UTCTime

envTokenScope :: ByteString

envUserAgent :: ByteString