Reddit.Example

Description

When run, this script will log each new comment posted in /r/haskell to standard output, scans it for the phrase "Haskell is great!!!!!!!!!", and if that is found, replies to them with "Indeed, it is!". (I put in 9 exclamation marks because I don't want to be morally responsible for somebody gratuitously importing it, running it, and spamming the sub.)

Anyway, you should look at the source code for this module instead of the built Haddocks. It contains many explanatory comments.

Documentation

getEnvAsText :: Text -> IO Text

getEnv but acts on Text values.

replyIfHaskellGreat :: Int -> Comment -> RedditT IO Int

An example of a callback function acting on comments. It takes an Int input state, the comment it sees, and does some kind of action on Reddit that ultimately returns another Int. The returned value will be used as the input the next time the callback is executed, i.e. on the next comment.

main :: IO ()

Run the bot.