Reddit

Types

type RedditT = ReaderT RedditEnv IO

The type of Reddit computations.

runRedditT :: RedditT a -> RedditEnv -> IO a

Run a Reddit computation using an env value obtained through authorisation (see Reddit.Auth).

runRedditT' :: RedditEnv -> RedditT a -> IO a

runRedditT' is flip runRedditT and is often more ergonomic.

Users

Fetch a user. TODO: Implement the right type.

user :: Text -> RedditT ByteString

Get information about a user

Posts

Fetch the first 25 posts from a given subreddit ordered by the SubredditSort parameter.

data Post

A single post

Constructors

Post

Fields

id' :: ID Post

url :: Text

author :: Text

title :: Text

body :: Text

subreddit :: Text

Empty if not a selftext post. TODO: sort out link posts

subreddit_id :: ID Subreddit

data Timeframe

Constructors

Hour

Day

Week

Month

Year

All

data SubredditSort

Constructors

Hot

New

Random

Rising

Top Timeframe

Controversial Timeframe

subredditPosts :: Text -> SubredditSort -> RedditT [Post]

Get the first 25 posts on a subreddit

Comments

Blah.

data Comment

A single comment.

Constructors

Comment

Fields

id' :: ID Comment

url :: Text

author :: Text

body :: Text

subreddit :: Text

subreddit_id :: ID Subreddit

post_id :: ID Post

parent_id :: Either (ID Comment) (ID Post)

addNewComment

Arguments

:: CanCommentOn a

=> ID a
The ID of the thing being replied to.

-> Text
The contents of the comment (in Markdown)

-> RedditT ()

Add a new comment as a reply to an existing post or comment.

subredditComments :: Text -> RedditT [Comment]

Get the most recent 25 comments on a subreddit. This endpoint is undocumented!

Streams

Streams are a popular feature in the Python PRAW library. They are reproduced here as they are useful for accomplishing a variety of bot-related tasks, usually iterating over a list of most recent posts / comments on a subreddit.

The most general function is stream. Here is a simple usage example, which logs each new comment posted in /r/haskell to standard output, scans it for the phrase "Haskell is great!", and if that is found, replies to them with "Indeed, it is!". It is assumed that you have obtained an env value from the authentication procedure (see Reddit.Auth).

postStream and commentStream are specialised versions which save a tiny bit of typing.

{-# LANGUAGE OverloadedStrings #-}
{-# LANGUAGE OverloadedDotSyntax #-}

main = do
  env <- ...

  let replyToGreatHaskell :: Int -> Comment -> RedditT Int
      replyToGreatHaskell count cmt = do
        let p = liftIO . T.putStrLn
        p ""
        p "Found new comment!"
        p $ "By    : /u/" <> cmt.author
        p $ "Link  : " <> cmt.url
        p $ "Text  : " <> cmt.body
        p $ T.pack ("Comments seen so far: " <> show (count + 1))

        let triggerText = "Haskell is great!"
        let replyText = "Indeed, it is!"
        when (triggerText `T.isInfixOf` cmt.body)
         (p "Replying to it..." >> addNewComment cmt.id' replyText)
        pure (count + 1)

  runRedditT' env $ do
    stream True replyToGreatHaskell 0 (subredditComments "haskell")

stream

Arguments

:: Eq a

=> Bool
Whether to ignore the things found in the first request. Since the first request is run when the stream is started, this essentially amounts to ignoring everything posted before the stream is started. You will most likely want this to be True.

-> (state -> a -> RedditT state)
A callback function to execute on all things found.

-> state
The initial state for the callback function.

-> RedditT [a]
The source of things to iterate over.

-> RedditT ()

If you have an action which generates a list of things (with the type RedditT [a]), then stream turns this an action which executes a callback function on an infinite list of things. It does so by repeatedly fetching the list of things (every five seconds). Here, 'things' refers to comments, posts, and so on.

Apart from the thing being acted on, the callback function is allowed to also take, as input, some kind of state, and update that state by returning a new value. This allows the user to, for example, keep track of how many things have been seen so far, or perform actions conditionally based on what the stream has previously thrown up.

stream'

Arguments

:: Eq a

=> Bool
Whether to ignore the first request.

-> (a -> RedditT ())
A callback function to execute on all things found.

-> RedditT [a]
The source of things to iterate over.

-> RedditT ()

stream' is a simpler version of stream, which accepts a callback that doesn't use state.

postStream

Arguments

:: Bool
Whether to ignore the first request.

-> (state -> Post -> RedditT state)
Callback function.

-> state
Initial state for callback.

-> Text
Subreddit name.

-> RedditT ()

Get a stream of new posts on a subreddit.

commentStream

Arguments

:: Bool
Whether to ignore the first request.

-> (state -> Comment -> RedditT state)
Callback function.

-> state
Initial state for callback.

-> Text
Subreddit name.

-> RedditT ()

Get a stream of new comments on a subreddit. Note that this also includes edited comments (that's not a design choice, it's just how the Reddit API works).