Reddit

Description

reddit-oauth2 provides a collection of functions to query the Reddit OAuth2 API.

It is currently very primitive, but contains enough stuff to let you make a bot that replies to certain phrases found in comments / posts (which is a very common use case). If you are interested in this specific use case, the fastest way to get up to scratch is probably to see the example in the Reddit.Example module.

Otherwise, the remainder of this page should be fairly well-documented and otherwise self-explanatory. Please feel free to raise an issue if there are any unclear aspects.

The RedditT monad

type RedditT = ReaderT RedditEnv IO

The type of a Reddit computation.

runRedditT :: RedditT a -> RedditEnv -> IO a

Run a Reddit computation using an env value obtained through authorisation (see Authentication).

runRedditT' :: RedditEnv -> RedditT a -> IO a

runRedditT' is flip runRedditT and is slightly more ergonomic.

data RedditEnv

Everything required to query the Reddit API.

This is a record type which includes things that you might want to change (e.g. configuration variables), as well as some things that you really should not change (e.g. the access token). Because of this, not all field accessors are exported; but due to a bug in Haddock, the ones that are exported are documented as separate functions instead of 'standalone' field accessors. However, they are still field accessors, so you can still do things like:

   env <- withCredentials ...
   let modifiedEnv = env {streamDelay = 2}

To be clear, the exported field accessors are: streamDelay only (so far).

streamDelay :: RedditEnv -> Double

Delay (in seconds) between successive requests when using streams. Reddit says you should not be querying more than 60 times in a minute, so this should not go below 1. Defaults to 5.

Authentication with account credentials

This section describes the credentials needed for using the Reddit API as a 'script'.

In order to obtain these, you will need to:

Log in using the account you want to post or query Reddit as (this may be a bot account, for example)

Navigate to https://www.reddit.com/prefs/apps

Click 'create another app' at the bottom and select the 'script' radio button

Fill in the required details. For a personal script, the redirect URI is not important, set it to anything you like.

Take note of the app ID (a string with random characters) and the secret (the same but longer). These become, respectively, clientID and clientSecret.

username and password are the Reddit login credentials of the account you are using.

Note: Do NOT ever share your password or the client secret publicly!

See also: https://github.com/reddit-archive/reddit/wiki/OAuth2-Quick-Start-Example

data Credentials

Constructors

Credentials

Fields

username :: Text

password :: Text

clientID :: Text

clientSecret :: Text

withCredentials

Arguments

:: Credentials

-> Text
Your user-agent. Reddit says you should use a unique and identifiable user-agent.

-> IO RedditEnv

Exchange user account credentials for a RedditEnv, which is required to run all Reddit queries.

Comments

Comments.

data Comment

A single comment.

Constructors

Comment

Fields

id' :: ID Comment

url :: Text

Permalink.

score :: Int

author :: Text

Username of the author (without the /u/).

author_id :: ID Account

body :: Text

Body text in Markdown.

subreddit :: Text

Name of the subreddit it was posted on (without the /r/ prefix).

subreddit_id :: ID Subreddit

post_id :: ID Post

is_submitter :: Bool

Whether the commenter also made the post.

parent_id :: Either (ID Comment) (ID Post)

If the comment is a top-level comment, this is the same as Right of post_id; otherwise it's Left of parent comment's ID.

createdTime :: UTCTime

editedTime :: EditedUTCTime

getComments :: [ID Comment] -> RedditT [Comment]

Fetch a list of comments by their IDs. More efficient than map getComment because it only makes one API call.

TODO: This uses the /api/info endpoint, which (apparently) provides slightly less information than /comments/<article>. For example, this endpoint doesn't provide a list of replies. This is probably worth investigating.

getComment :: ID Comment -> RedditT Comment

Fetch a single comment given its ID.

addNewComment

Arguments

:: CanCommentOn a

=> ID a
The ID of the thing being replied to.

-> Text
The contents of the comment (in Markdown)

-> RedditT ()

Add a new comment as a reply to an existing post or comment.

accountComments

Arguments

:: Int
Number of comments to fetch. See the listings section for an explanation of this parameter.

-> Text
Username (without the /u/).

-> RedditT [Comment]

Get the most recent comments by a user.

subredditComments

Arguments

:: Int
Number of comments to fetch. See the listings section for an explanation of this parameter.

-> Text
Subreddit name (without the /r/).

-> RedditT [Comment]

Get the most recent comments on a subreddit.

Accounts (i.e. users)

Accounts.

data Account

Constructors

Account

Fields

id' :: ID Account

username :: Text

Username (without the /u/ prefix).

display_name :: Text

The name that's shown on their profile. Empty if not set.

link_karma :: Int

comment_karma :: Int

total_karma :: Int

createdTime :: UTCTime

getAccounts :: [ID Account] -> RedditT [Account]

Fetch a list of accounts given their IDs. More efficient than map getAccount because it only makes one API call.

getAccount :: ID Account -> RedditT Account

Fetch a single account given its ID.

getAccountByName

Arguments

:: Text
Username (without the /u/).

-> RedditT Account

Fetch details about a named user.

Posts

Fetch the first 25 posts from a given subreddit ordered by the SubredditSort parameter.

data Post

A single post

Constructors

Post

Fields

id' :: ID Post

url :: Text

Permalink.

score :: Int

author :: Text

Username of the author (without the /u/).

author_id :: ID Account

title :: Text

body :: Text

Empty string if not a text post.

content_url :: Text

For a link post, this is the link. For a text post, this is the same as url.

flairtext :: Maybe Text

None if not flaired.

subreddit :: Text

Name of the subreddit (without the /r/)

subreddit_id :: ID Subreddit

createdTime :: UTCTime

editedTime :: EditedUTCTime

getPosts :: [ID Post] -> RedditT [Post]

Fetch a list of posts by their IDs. More efficient than map getPost because it only makes one API call.

getPost :: ID Post -> RedditT Post

Fetch a single post given its ID.

accountPosts

Arguments

:: Int
Number of posts to fetch. See the listings section for an explanation of this parameter.

-> Text
Username (without the /u/).

-> RedditT [Post]

Get the most recent posts by a user.

data Timeframe

Constructors

Hour

Day

Week

Month

Year

All

data SubredditSort

Constructors

Hot

New

Random

Rising

Top Timeframe

Controversial Timeframe

subredditPosts

Arguments

:: Int
Number of posts to fetch. See the listings section for an explanation of this parameter.

-> Text
Subreddit name (without the /r/).

-> SubredditSort
Sort type for the subreddit posts. Entirely analogous to the options when browsing Reddit on the web.

-> RedditT [Post]

Get the posts from the front page of a subreddit.

Messages

Messages aren't implemented yet.

data Message

Constructors

Message

Fields

id' :: ID Message

Subreddits

Subreddits.

data Subreddit

Constructors

Subreddit

Fields

id' :: ID Subreddit

name :: Text

Subreddit name (without the /r/).

title :: Text

Human name (the string that appears at the top of the sub).

description :: Text

Sidebar text.

subscribers :: Int

Number of subscribers.

created :: UTCTime

Created time.

getSubreddits :: [ID Subreddit] -> RedditT [Subreddit]

Fetch a list of subreddits by their IDs. More efficient than map getSubreddit because it only makes one API call.

getSubreddit :: ID Subreddit -> RedditT Subreddit

Fetch a single subreddit given its ID.

getSubredditsByName

Arguments

:: [Text]
List of subreddit names (without the /r/'s).

-> RedditT [Subreddit]

Fetch a list of subreddits by their names. More efficient than map getSubredditByName.

getSubredditByName

Arguments

:: Text
Subreddit name (without the /r/).

-> RedditT Subreddit

Fetch a single subreddit by its ID.

Awards

Awards aren't implemented yet.

data Award

Constructors

Award

Fields

id' :: ID Award

Streams

Streams are a popular feature in the Python PRAW library. They are reproduced here as they are useful for accomplishing a variety of bot-related tasks, usually iterating over a list of most recent posts / comments on a subreddit.

The most general function is stream; a usage example is provided at the top of this file. postStream and commentStream are specialised versions which save a tiny bit of typing.

stream

Arguments

:: Eq a

=> (state -> a -> RedditT state)
A callback function to execute on all things found.

-> state
The initial state for the callback function.

-> RedditT [a]
The source of things to iterate over.

-> RedditT ()

If you have an action which generates a list of things (with the type RedditT [a]), then "stream" turns this an action which executes a callback function on an infinite list of things. It does so by repeatedly fetching the list of either comments or posts.

Apart from the thing being acted on, the callback function is allowed to also take, as input, some kind of state, and update that state by returning a new value. This allows the user to, for example, keep track of how many things have been seen so far, or perform actions conditionally based on what the stream has previously thrown up.

You can adjust the frequency with which the stream is refreshed, and secondly, the number of items fetched on each request. These can be changed using the streamDelay field of the RedditEnv you are using. The local function is particularly helpful here.

stream'

Arguments

:: Eq a

=> (a -> RedditT ())
Whether to ignore the first request.

-> RedditT [a]
The source of things to iterate over.

-> RedditT ()

stream' is a simpler version of stream, which accepts a callback that doesn't use state.

postStream

Arguments

:: (state -> Post -> RedditT state)
Callback function.

-> state
Initial state for callback.

-> Text
Subreddit name (without the /r/).

-> RedditT ()

Get a stream of new posts on a subreddit.

commentStream

Arguments

:: (state -> Comment -> RedditT state)
Callback function.

-> state
Initial state for callback.

-> Text
Subreddit name (without the /r/).

-> RedditT ()

Get a stream of new comments on a subreddit. Note that this also includes edited comments (that's not a design choice, it's just how the Reddit API works).

Listings

Listings are Reddit's way of paginating things (i.e. comments, posts, etc.). When you ask for a series of things, the Reddit API really returns a listing of things, from which the posts themselves must be extracted.

This library abstracts this away from you completely, so you should not have to deal with any listings themselves. However, there's a good reason to know about how listings work, namely efficiency.

When making a single HTTP request to the Reddit API, only 100 things can be returned in a single listing. So, if you want to fetch the first 200 comments of a user, with accountComments 200 "username", then we need to perform two HTTP requests.

This goes up to a maximum of 1000, after which Reddit refuses to return any more results. So, even if you ask for 1500 comments, you'll only get the first 1000. Also, if you ask for 1000 comments and the user has only ever commented 500 times, you'll get all 500.

This leads us to three guidelines for choosing the Int parameter for functions such as accountComments, accountPosts, etc.:

Keep it to 100 or lower if you want it to complete within a single request. In particular, you probably don't want to make this greater than 100 if you're creating a stream.

If you just want 'as many as possible', put 1000.

Don't bother putting anything larger than 1000.

Other types

data ID a where

Reddit 'things' have an ID, which (in the raw JSON) is a string prefixed by "t1_", "t2_", "t3_"... depending on what type of object it is. In this library, IDs are stored without this prefix; so CommentID "abcde" is equivalent to Reddit's "t1_abcde".

By the way, this is my first time dabbling with GADTs in a library, so I'm not completely sure if it's the correct (or the smartest) way to do it. Please get in touch if you want to tell me about alternatives.

Constructors

CommentID :: Text -> ID Comment
t1_

AccountID :: Text -> ID Account
t2_

PostID :: Text -> ID Post
t3_

MessageID :: Text -> ID Message
t4_

SubredditID :: Text -> ID Subreddit
t5_

AwardID :: Text -> ID Award
t6_

class HasID a where

Things which have an ID.

Methods

mkFullName :: a -> Text

Construct the full name of a thing from the thing itself.

mkFullNameFromID :: ID a -> Text

Construct the full name of a thing from its ID. This is possible because the ID itself carries information about its type, allowing us to figure out the correct prefix.

class HasID a => CanCommentOn a

Things that can be commented on (i.e. comments and posts).

data EditedUTCTime

Reddit's API does not always preserve the time that a comment was edited (particularly for old comments, I think), so these represent the three possibilities for the editedTime field of comments and posts.

Constructors

NeverEdited

EditedButTimeUnknown

EditedAt UTCTime