Reddit

Description

reddit-oauth2 provides a collection of functions to query the Reddit OAuth2 API.

It is currently very primitive, but contains enough stuff to let you make a bot that replies to certain phrases found in comments / posts (which is a very common use case).

It is probably easiest to demonstrate this with an example. One such example is provided in srcRedditExample.hs. When run, this script will log each new comment posted in /r/haskell to standard output, scans it for the phrase "Haskell is great!!!!!!!!!", and if that is found, replies to them with "Indeed, it is!".

(Even though I don't think many people post that exact phrase on /r/haskell, I put in 9 exclamation marks because I don't want to be morally responsible for somebody gratuitously importing it.)

Types

data RedditEnv

Everything required to query the Reddit API.

type RedditT = ReaderT RedditEnv IO

The type of a Reddit computation.

runRedditT :: RedditT a -> RedditEnv -> IO a

Run a Reddit computation using an env value obtained through authorisation (see Reddit.Auth).

runRedditT' :: RedditEnv -> RedditT a -> IO a

runRedditT' is flip runRedditT and is often more ergonomic.

Authentication with account credentials

This section describes the credentials needed for using the Reddit API as a 'script'.

In order to obtain these, you will need to:

Log in using the account you want to post or query Reddit as (this may be a bot account, for example)

Navigate to https://www.reddit.com/prefs/apps

Click 'create another app' at the bottom and select the 'script' radio button

Fill in the required details. For a personal script, the redirect URI is not important, set it to anything you like.

Take note of the app ID (a string with random characters) and the secret (the same but longer). These become, respectively, clientID and clientSecret.

username and password are the Reddit login credentials of the account you are using.

Note: Do NOT ever share your password or the client secret publicly!

See also: https://github.com/reddit-archive/reddit/wiki/OAuth2-Quick-Start-Example

data Credentials

Constructors

Credentials

Fields

username :: Text

password :: Text

clientID :: Text

clientSecret :: Text

withCredentials

Arguments

:: Credentials

-> Text
Your user-agent. Reddit says you should use a unique and identifiable user-agent.

-> IO RedditEnv

Exchange user account credentials for a RedditEnv, which is required to run all Reddit queries.

Users

Fetch a user. TODO: Implement the right type.

user :: Text -> RedditT ByteString

Get information about a user

Posts

Fetch the first 25 posts from a given subreddit ordered by the SubredditSort parameter.

data Post

A single post

Constructors

Post

Fields

id' :: ID Post

url :: Text

author :: Text

title :: Text

body :: Text

subreddit :: Text

Empty if not a selftext post. TODO: sort out link posts

subreddit_id :: ID Subreddit

data Timeframe

Constructors

Hour

Day

Week

Month

Year

All

data SubredditSort

Constructors

Hot

New

Random

Rising

Top Timeframe

Controversial Timeframe

subredditPosts :: Text -> SubredditSort -> RedditT [Post]

Get the first 25 posts on a subreddit

Comments

Blah.

data Comment

A single comment.

Constructors

Comment

Fields

id' :: ID Comment

url :: Text

author :: Text

body :: Text

subreddit :: Text

subreddit_id :: ID Subreddit

post_id :: ID Post

parent_id :: Either (ID Comment) (ID Post)

addNewComment

Arguments

:: CanCommentOn a

=> ID a
The ID of the thing being replied to.

-> Text
The contents of the comment (in Markdown)

-> RedditT ()

Add a new comment as a reply to an existing post or comment.

subredditComments :: Text -> RedditT [Comment]

Get the most recent 25 comments on a subreddit. This endpoint is undocumented!

Streams

Streams are a popular feature in the Python PRAW library. They are reproduced here as they are useful for accomplishing a variety of bot-related tasks, usually iterating over a list of most recent posts / comments on a subreddit.

The most general function is stream; a usage example is provided at the top of this file. postStream and commentStream are specialised versions which save a tiny bit of typing.

stream

Arguments

:: Eq a

=> Bool
Whether to ignore the things found in the first request. Since the first request is run when the stream is started, this essentially amounts to ignoring everything posted before the stream is started. You will most likely want this to be True.

-> (state -> a -> RedditT state)
A callback function to execute on all things found.

-> state
The initial state for the callback function.

-> RedditT [a]
The source of things to iterate over.

-> RedditT ()

If you have an action which generates a list of things (with the type RedditT [a]), then stream turns this an action which executes a callback function on an infinite list of things. It does so by repeatedly fetching the list of things (every five seconds). Here, 'things' refers to comments, posts, and so on.

Apart from the thing being acted on, the callback function is allowed to also take, as input, some kind of state, and update that state by returning a new value. This allows the user to, for example, keep track of how many things have been seen so far, or perform actions conditionally based on what the stream has previously thrown up.

stream'

Arguments

:: Eq a

=> Bool
Whether to ignore the first request.

-> (a -> RedditT ())
A callback function to execute on all things found.

-> RedditT [a]
The source of things to iterate over.

-> RedditT ()

stream' is a simpler version of stream, which accepts a callback that doesn't use state.

postStream

Arguments

:: Bool
Whether to ignore the first request.

-> (state -> Post -> RedditT state)
Callback function.

-> state
Initial state for callback.

-> Text
Subreddit name.

-> RedditT ()

Get a stream of new posts on a subreddit.

commentStream

Arguments

:: Bool
Whether to ignore the first request.

-> (state -> Comment -> RedditT state)
Callback function.

-> state
Initial state for callback.

-> Text
Subreddit name.

-> RedditT ()

Get a stream of new comments on a subreddit. Note that this also includes edited comments (that's not a design choice, it's just how the Reddit API works).