Reddit

Description

reddit-oauth2 provides a collection of functions to query the Reddit OAuth2 API.

It is currently very primitive, but contains enough stuff to let you make a bot that replies to certain phrases found in comments / posts (which is a very common use case). If you are interested in this specific use case, the fastest way to get up to scratch is probably to see the example in the Reddit.Example module.

Otherwise, the remainder of this page should be fairly well-documented and otherwise self-explanatory. Please feel free to raise an issue if there are any unclear aspects.

The RedditT monad

type RedditT = ReaderT RedditEnv IO

The type of a Reddit computation.

data RedditEnv

Everything required to query the Reddit API.

This is a record type. However, in principle, you should not need to edit anything here, so the field accessors are not exported. If you need to extract the token, you can use getTokenFromEnv.

runRedditT :: RedditT a -> RedditEnv -> IO a

Run a Reddit computation using a RedditEnv value obtained through authentication (see Authentication).

runRedditT' :: RedditEnv -> RedditT a -> IO a

runRedditT' is flip runRedditT and is slightly more ergonomic.

runRedditTCleanup :: RedditT a -> RedditEnv -> IO a

Run a Reddit computation, and additionally clean up the RedditEnv value by revoking the active access token after the computation finishes.

Doing this is generally good behaviour. Alternatively, you can manually use revokeToken.

runRedditTCleanup' :: RedditEnv -> RedditT a -> IO a

Same as flip runRedditTCleanup.

revokeToken :: RedditEnv -> IO ()

Revoke an access token contained in a RedditEnv, rendering it unusable. If you want to continue performing queries after this, you will need to generate a new RedditEnv.

Manual environment management

For more complicated use cases, you will almost certainly have to perform more manual management of the RedditEnv value. Specifically, you will want to extract tokens from envs (for example, to store them in a database), and re-generate new envs from existing tokens.

The functions in this section aim to make this process smoother.

getTokenFromEnv :: RedditT Token

Extract the current OAuth2 token being used.

mkEnvFromToken

Arguments

:: Token

-> Text
User-agent

-> IO RedditEnv

Yada.

Authentication

The first step to querying the Reddit API is to authenticate with Reddit. (Technically, you don't have to; however, you will be subjected to a much lower rate limit. This library does not support unauthenticated usage of the Reddit API.)

The method in which you authenticate will depend on the type of application you are building.

Most Reddit bots, or scripts, only require you to authenticate as a single user (which may be a bot account). In this case, you can simply log in with your account credentials, using the OwnerCredentials constructor. If you are not sure what you are doing, this is probably what you want.

If you are instead building an application which is meant to be used by other people (e.g. a web app to let visitors analyse their posts), then you need to authenticate via the 'code grant' method, using the CodeGrantCredentials constructor.

In either case, you need to construct the appropriate Credentials and pass them to the authenticate function, which will give you a RedditEnv value that you can then use to query Reddit.

data Credentials

The Credentials type contains all the information you need to authenticate with the Reddit API.

Constructors

OwnerCredentials
'Resource owner' flow

Formally, this is the 'Resource Owner Password Credentials Grant' for OAuth 2.0, which is described in Section 4.3 of IETF RFC 6749.

To authenticate in this manner, you will need a username and password for a Reddit account, plus a client ID and client secret associated with a Reddit app. The username in question must be listed as a developer of the app. You can set up a new app by:

Log in using the account you want to post or query Reddit as (this may be a bot account, for example).

Navigate to https://www.reddit.com/prefs/apps

Click 'create another app' at the bottom and select the 'script' radio button

Fill in the required details. For a personal script, the redirect URI is not important, set it to anything you like.

Take note of the app ID (a string with random characters) and the secret (the same but longer). These become, respectively, clientID and clientSecret.

username and password are the Reddit login credentials of the account you are using.

Note: Do NOT ever share your password or the client secret publicly!

See also: https://github.com/reddit-archive/reddit/wiki/OAuth2-Quick-Start-Example

Fields

ownerUsername :: Text

ownerPassword :: Text

ownerClientId :: Text

ownerClientSecret :: Text

CodeGrantCredentials
'Code grant' flow

Formally, this is the 'Authorisation Code Grant' for OAuth 2.0, which is described in Section 4.1 of IETF RFC 6749.

Begin by setting up a new Reddit app in the same way as for the 'resource owner' flow above. Then:

codeGrantClientId and codeGrantClientSecret are simply the 'app ID' and 'app secret' from the Reddit app.

codeGrantRedirectUri is the redirect URI you specified when setting up the app. Note that this must match exactly.

To obtain codeGrantCode, you must first construct the Reddit authentication link using mkRedditAuthURL. Send your user to this link, and they will be greeted with a Reddit page asking them to authorise your app. When they click 'allow', Reddit will send the user back to the redirect URI which you specified, along with a 'code' query parameter in the URI. You should then parse this and set it as codeGrantCode.

See also: https://github.com/reddit-archive/reddit/wiki/OAuth2

Fields

codeGrantClientId :: Text

codeGrantClientSecret :: Text

codeGrantRedirectUri :: Text

codeGrantCode :: Text

NoCredentials
Do not provide credentials. This is used when 'reconstructing' a RedditEnv from an existing token (e.g. one stored in a database) using mkEnvFromToken. You should not need to use this.

authenticate

Arguments

:: Credentials

-> Text
Your user-agent. Reddit says you should use a unique and identifiable user-agent.

-> IO RedditEnv

Once you have set up your Credentials, you can use this function to exchange them for a token contained inside a RedditEnv, which can then be used to perform all Reddit queries.

data AuthUrlParams

The information required for your app to construct a link for the user to 'authorise with Reddit'.

This is only required for the 'code grant' flow.

Constructors

AuthUrlParams

Fields

authUrlClientID :: Text

This is the client ID for your app.

authUrlState :: Maybe Text

This can be any text value you want. If the user is successfully authorised, they will be sent back to your redirect URI, and the same state will be included as a query parameter.

authUrlRedirectUri :: Text

The URI which users will be sent back to after logging in. This must match the redirect URI set up in https://reddit.com/prefs/apps.

authUrlDuration :: Duration

Temporary or Permanent, depending on how long you need your user to be logged in for. Temporary tokens last for 1 hour, after which the user must re-authenticate if they want to continue using your app.

authUrlScopes :: Set Scope

The OAuth2 scopes that you need. This depends on what your app does on behalf of its users. See Scope for the full list.

mkRedditAuthURL

Arguments

:: AuthUrlParams

-> Bool
Whether to use the mobile version of the authorisation page

-> Text
The authorisation URL

Construct the Reddit authorisation URL that you should send your users to for the 'code grant' flow.

data Scope

Scopes are used to limit the things which a token can do. You should ensure that you only request a token for the scopes that you need. When authenticating with Reddit, the user will be able to see which scopes are being requested.

To see which scope(s) are required by each endpoint, see https://www.reddit.com/dev/api/oauth. Note that this library implements only a small subset of endpoints, so most of these are not very useful right now.

You can obtain this list of scopes via https://reddit.com/api/v1/scopes.

Constructors

ScopeCreddits
Spend the logged in user's Reddit gold.

ScopeModNote
Access mod notes for subreddits moderated by the logged in user.

ScopeModContributors
Add/remove users to/from approved user lists, ban/unban, and mute/unmute users on subreddits moderated by the logged in user.

ScopeModMail
Access and manage modmail for subreddits moderated by the logged in user.

ScopeModConfig
Manage the configuration, sidebar, and CSS for subreddits moderated by the logged in user.

ScopeSubscribe
Manage subreddit subscriptions as well as 'friends'.

ScopeStructuredStyles
Edit structured styles for subreddits moderated by the logged in user.

ScopeVote
Vote on posts and comments.

ScopeWikiEdit
Edit wiki pages.

ScopeMySubreddits
Access the list of subreddits moderated, contributed to, and subscribed to.

ScopeSubmit
Submit links and comments.

ScopeModLog
Access the moderation log in subreddits moderated by the logged in user.

ScopeModPosts
Approve, remove, mark as (N)SFW, and distinguish content in subreddits moderated by the logged in user.

ScopeModFlair
Manage and assign flair in subreddits moderated by the logged in user.

ScopeSave
Save and unsave posts and comments.

ScopeModOthers
Invite and remove moderators on subreddits moderated by the logged in user.

ScopeRead
Access posts and comments.

ScopePrivateMessages
Read and send private messages. (Note: this does not include chat. There is, at present, no way to access chat via the public Reddit API.)

ScopeReport
Report, hide, and unhide content.

ScopeIdentity
Get Reddit username and signup date.

ScopeLiveManage
Manage settings and contributors of live threads.

ScopeAccount
Update preferences and other account information. Does not grant access to email or password.

ScopeModTraffic
Access traffic stats in subreddits moderated by the user.

ScopeWikiRead
Read wiki pages.

ScopeEdit
Edit and delete posts and comments.

ScopeModWiki
Change editors and visibility of wiki pages in subreddits moderated by the logged in user.

ScopeModSelf
Accept invitations to moderate a subreddit and step down from moderating subreddits.

ScopeHistory
Access voting history, saved submissions, and hidden submissions.

ScopeFlair
Edit subreddit and post flair.

allScopes :: Set Scope

A quick way to get all scopes.

data Duration

This should be chosen depending on whether your app needs a temporary token (i.e. user must authenticate again after the token expires), or a permanent one (i.e. user does not need to authenticate again).

Internally, permanent tokens are handled by requesting a refresh token when the existing one expires.

Constructors

Temporary

Permanent

Comments

Comments.

data Comment

A single comment.

Note that the Eq and Ord instances for Comment only compare the comment ID.

Constructors

Comment

Fields

commentId :: ID Comment

commentUrl :: Text

Permalink.

commentScore :: Int

commentAuthor :: Text

Username of the author (without the /u/).

commentAuthorId :: Maybe (ID Account)

Can be Nothing if the account was deleted.

commentBody :: Text

Body text in Markdown.

commentSubreddit :: Text

Name of the subreddit it was posted on (without the /r/ prefix).

commentSubredditId :: ID Subreddit

commentPostId :: ID Post

commentIsSubmitter :: Bool

Whether the commenter also made the post.

commentParentId :: Either (ID Comment) (ID Post)

If the comment is a top-level comment, this is the same as Right of post_id; otherwise it's Left of parent comment's ID.

commentCreatedTime :: UTCTime

commentEditedTime :: EditedUTCTime

data CommentTree

When fetching comments on a specific post, we need a tree-like structure to account for the relationship between comments.

Each post is associated with a list of top-level CommentTrees. Each CommentTree is either a comment together with its replies, or a special MoreComments value which represents the fact that a discussion continues further, but the data were not returned immediately by Reddit.

Constructors

ActualComment Comment [CommentTree]

MoreComments [ID Comment]

getComments :: [ID Comment] -> RedditT [Comment]

Fetch a list of comments by their IDs. More efficient than map getComment because it only makes one API call.

This function can work with any collection of comments (they may, for example, be scattered amongst a whole range of posts). However, it does not return a list of replies for each comment. If you want to see comment replies, you need to use getPostWithComments, which internally uses a different API endpoint.

getComment :: ID Comment -> RedditT Comment

Fetch a single comment given its ID. The same caveat described for getComments also applies here.

addNewComment

Arguments

:: CanCommentOn a

=> ID a
The ID of the thing being replied to.

-> Text
The contents of the comment (in Markdown)

-> RedditT ()

Add a new comment as a reply to an existing post or comment.

accountComments

Arguments

:: Int
Number of comments to fetch (maximum 1000). See the listings section for an explanation of this parameter.

-> Text
Username (without the /u/).

-> RedditT [Comment]

Get the most recent comments by a user.

subredditComments

Arguments

:: Int
Number of comments to fetch (maximum 1000). See the listings section for an explanation of this parameter.

-> Text
Subreddit name (without the /r/).

-> RedditT [Comment]

Get the most recent comments on a subreddit.

getPostWithComments :: ID Post -> RedditT (Post, [CommentTree])

Retrieve a post, together with a tree of comments on it.

This comment tree does not necessarily contain all comments on the post. For posts with many comments, Reddit often hides comments which are above a certain depth / very low in the thread. This can be seen on the website, where you need to click 'X more replies' to load the comments fully. In this library, this is modelled as a MoreComments value.

The Reddit API does not provide any way to directly obtain these collapsed replies: instead, they must be expanded one by one. You can use either expandTree to perform expansion step-by-step, or use expandTreeFully to recursively expand every collapsed reply.

expandTree :: ID Post -> [CommentTree] -> RedditT [CommentTree]

Expand the first instance of MoreComments in a tree. The required post ID is that of the post for which you are expanding comments.

expandTreeFully :: ID Post -> [CommentTree] -> RedditT [CommentTree]

Expand all instances of MoreComments in a tree (including those unveiled by expanding previous MoreComments).

Essentially, if you want to obtain every single comment on a post given its ID (pid :: ID Post), then you can do:

(_, trees) <- getPostWithComments pid
expandedTrees <- expandTreeFully pid trees

Note that this can take quite a bit of time to run for posts with thousands of comments. It can't be sped up either, because Reddit's API stipulates that you cannot make concurrent requests to this endpoint.

Accounts (i.e. users)

Accounts.

data Account

Constructors

Account

Fields

accountId :: ID Account

accountUsername :: Text

Username (without the /u/ prefix).

accountDisplayName :: Text

The name that's shown on their profile. Empty if not set.

accountLinkKarma :: Int

accountCommentKarma :: Int

accountTotalKarma :: Int

accountCreatedTime :: UTCTime

myAccount :: RedditT Account

Fetch the account of the currently logged-in user.

getAccounts :: [ID Account] -> RedditT [Account]

Fetch a list of accounts given their IDs. More efficient than map getAccount because it only makes one API call.

getAccount :: ID Account -> RedditT Account

Fetch a single account given its ID.

getAccountByName

Arguments

:: Text
Username (without the /u/).

-> RedditT Account

Fetch details about a named user.

Posts

data Post

A single post.

Note that the Eq and Ord instances for Post only compare the post ID.

Constructors

Post

Fields

postId :: ID Post

postUrl :: Text

Permalink.

postScore :: Int

postAuthor :: Text

Username of the author (without the /u/).

postAuthorId :: Maybe (ID Account)

Can be Nothing if the account was deleted.

postTitle :: Text

postBody :: Text

Empty string if not a text post.

postContentUrl :: Text

For a link post, this is the link. For a text post, this is the same as url.

postFlairText :: Maybe Text

None if not flaired.

postSubreddit :: Text

Name of the subreddit (without the /r/)

postSubredditId :: ID Subreddit

postCreatedTime :: UTCTime

postEditedTime :: EditedUTCTime

getPosts :: [ID Post] -> RedditT [Post]

Fetch a list of posts by their IDs. More efficient than map getPost because it only makes one API call.

getPost :: ID Post -> RedditT Post

Fetch a single post given its ID.

If you want to fetch a post together with its comments, you can use getPostWithComments.

accountPosts

Arguments

:: Int
Number of posts to fetch (maximum 1000). See the listings section for an explanation of this parameter.

-> Text
Username (without the /u/).

-> RedditT [Post]

Get the most recent posts by a user.

data Timeframe

The timeframe over which to get top or most controversial posts.

Constructors

Hour

Day

Week

Month

Year

All

data SubredditSort

Sorting order for posts in a subreddit.

Constructors

Hot

New

Random

Rising

Top Timeframe

Controversial Timeframe

subredditPosts

Arguments

:: Int
Number of posts to fetch (maximum 1000). See the listings section for an explanation of this parameter.

-> Text
Subreddit name (without the /r/).

-> SubredditSort
Sort type for the subreddit posts. Entirely analogous to the options when browsing Reddit on the web.

-> RedditT [Post]

Get the posts from the front page of a subreddit, with the ordering specified in the SubredditSort argument.

edit :: CanCommentOn a => ID a -> Text -> RedditT ()

Edit the contents of a comment or a text post. You must be authenticated as the person who posted it.

delete :: CanCommentOn a => ID a -> RedditT ()

Delete a post or a comment. You must be authenticated as the person who posted it.

If you want to remove a post or a comment on a subreddit you moderate, use remove instead.

remove

Arguments

:: CanCommentOn a

=> ID a

-> Bool
Whether the thing being removed is spam.

-> RedditT ()

As a moderator, remove a post or a comment. This is the inverse of approve.

approve :: CanCommentOn a => ID a -> RedditT ()

As a moderator, approve a post or a comment. This is the inverse of remove.

Messages

Messages aren't implemented yet.

data Message

Constructors

Message

Fields

messageId :: ID Message

Subreddits

Subreddits.

data Subreddit

Constructors

Subreddit

Fields

subredditId :: ID Subreddit

subredditName :: Text

Subreddit name (without the /r/).

subredditTitle :: Text

Human name (the string that appears at the top of the sub).

subredditDescription :: Text

Sidebar text.

subredditSubscribers :: Int

Number of subscribers.

subredditCreated :: UTCTime

Created time.

getSubreddits :: [ID Subreddit] -> RedditT [Subreddit]

Fetch a list of subreddits by their IDs. More efficient than map getSubreddit because it only makes one API call.

getSubreddit :: ID Subreddit -> RedditT Subreddit

Fetch a single subreddit given its ID.

getSubredditsByName

Arguments

:: [Text]
List of subreddit names (without the /r/'s).

-> RedditT [Subreddit]

Fetch a list of subreddits by their names. More efficient than map getSubredditByName.

getSubredditByName

Arguments

:: Text
Subreddit name (without the /r/).

-> RedditT Subreddit

Fetch a single subreddit by its ID.

Awards

Awards aren't implemented yet.

data Award

Constructors

Award

Fields

awardId :: ID Award

Streams

Streams are a popular feature in the Python PRAW library. They are reproduced here as they are useful for accomplishing a variety of bot-related tasks, usually iterating over a list of most recent posts / comments on a subreddit.

The most general function is stream; a usage example is provided in the Example module. postStream and commentStream are specialised versions which save a tiny bit of typing.

data StreamSettings

Available configuration variables for streams.

Constructors

StreamSettings

Fields

streamsDelay :: Double

Delay (in seconds) between successive requests when using streams. Reddit says you should not be querying more than 60 times in a minute, so this should not go below 1. Defaults to 5.

streamsStorageSize :: Int

Number of 'seen' items to keep in memory when running a stream. If you are requesting N items at a go, there doesn't appear to be much point in making this larger than N. N should probably be 100, but this defaults to 250 to be safe, because I'm not sure if there are weird edge cases.

defaultStreamSettings :: StreamSettings

Default stream settings. See StreamSettings for the specification of these values.

stream

Arguments

:: Eq a

=> StreamSettings
Settings.

-> (state -> a -> RedditT state)
A callback function to execute on all things found.

-> state
The initial state for the callback function.

-> RedditT [a]
The source of things to iterate over.

-> RedditT ()

If you have an action which generates a list of things (with the type RedditT [a]), then "stream" turns this an action which executes a callback function on an infinite list of things. It does so by repeatedly fetching the list of either comments or posts.

Apart from the thing being acted on, the callback function is allowed to also take, as input, some kind of state, and update that state by returning a new value. This allows the user to, for example, keep track of how many things have been seen so far, or perform actions conditionally based on what the stream has previously thrown up.

You can adjust the frequency with which the stream is refreshed, and secondly, the number of 'seen' items which are stored in memory. These can be changed using the delay and storageSize fields in the StreamSettings used.

stream'

Arguments

:: Eq a

=> StreamSettings
Settings.

-> (a -> RedditT ())
A simple callback function which iterates over all things seen.

-> RedditT [a]
The source of things to iterate over.

-> RedditT ()

stream' is a simpler version of stream, which accepts a callback that doesn't use state.

postStream

Arguments

:: StreamSettings
Settings.

-> (state -> Post -> RedditT state)
Callback function.

-> state
Initial state for callback.

-> Text
Subreddit name (without the /r/).

-> RedditT ()

Get a stream of new posts on a subreddit.

commentStream

Arguments

:: StreamSettings
Settings.

-> (state -> Comment -> RedditT state)
Callback function.

-> state
Initial state for callback.

-> Text
Subreddit name (without the /r/).

-> RedditT ()

Get a stream of new comments on a subreddit.

Note that this also includes edited comments (that's not a design choice, it's just how the Reddit API works). If you want to only act on newly posted comments, you can check the editedTime field of the comment in your callback function; however, note that this will still show up as NeverEdited for comments edited within the 3-minute grace period (i.e. 'ninja edits').

Listings

Listings are Reddit's way of paginating things (i.e. comments, posts, etc.). When you ask for a series of things, the Reddit API really returns a listing of things, from which the posts themselves must be extracted.

This library abstracts this away from you completely, so you should not have to deal with any listings themselves. However, there's a good reason to know about how listings work, namely efficiency.

When making a single HTTP request to the Reddit API, only 100 things can be returned in a single listing. So, if you want to fetch the first 200 comments of a user, with accountComments 200 "username", then we need to perform two HTTP requests. Of course, this library does that for you automatically—but it will take longer than you might initially expect.

This goes up to a maximum of 1000, after which Reddit refuses to return any more results. So, even if you ask for 1500 comments, you'll only get the first 1000. Also, if you ask for 1000 comments and the user has only ever commented 500 times, you'll get all 500.

This leads us to three guidelines for choosing the Int parameter for functions such as accountComments, accountPosts, etc.:

Keep it to 100 or lower if you want it to complete within a single request. In particular, you probably don't want to make this greater than 100 if you're creating a stream.

If you just want 'as many as possible', put 1000.

Don't bother putting anything larger than 1000.

Other types

data ID a where

Reddit 'things' have an ID, which (in the raw JSON) is a string prefixed by "t1_", "t2_", "t3_"... depending on what type of object it is. In this library, IDs are stored without this prefix; so CommentID "abcde" is equivalent to Reddit's "t1_abcde".

By the way, this is my first time dabbling with GADTs in a library, so I'm not completely sure if it's the correct (or the smartest) way to do it. Please get in touch if you want to tell me about alternatives.

Constructors

CommentID
t1_

Fields

:: { unCommentID :: Text

} -> ID Comment

AccountID
t2_

Fields

:: { unAccountID :: Text

} -> ID Account

PostID
t3_

Fields

:: { unPostID :: Text

} -> ID Post

MessageID
t4_

Fields

:: { unMessageID :: Text

} -> ID Message

SubredditID
t5_

Fields

:: { unSubredditID :: Text

} -> ID Subreddit

AwardID
t6_

Fields

:: { unAwardID :: Text

} -> ID Award

class HasID a where

Things which have an ID.

Methods

mkFullName :: a -> Text

Construct the full name of a thing from the thing itself.

mkFullNameFromID :: ID a -> Text

Construct the full name of a thing from its ID. This is possible because the ID itself carries information about its type, allowing us to figure out the correct prefix.

class HasID a => CanCommentOn a

Things that can be commented on (i.e. comments and posts).

data EditedUTCTime

Reddit's API does not always preserve the time that a comment was edited (particularly for old comments, I think), so these represent the three possibilities for the editedTime field of comments and posts.

Constructors

NeverEdited

EditedButTimeUnknown

EditedAt UTCTime