Reddit.Auth

Description

Documentation

data Credentials

The Credentials type contains all the information you need to authenticate with the Reddit API.

Constructors

OwnerCredentials
'Resource owner' flow

Formally, this is the 'Resource Owner Password Credentials Grant' for OAuth 2.0, which is described in Section 4.3 of IETF RFC 6749.

To authenticate in this manner, you will need a username and password for a Reddit account, plus a client ID and client secret associated with a Reddit app. The username in question must be listed as a developer of the app. You can set up a new app by:

Log in using the account you want to post or query Reddit as (this may be a bot account, for example).

Navigate to https://www.reddit.com/prefs/apps

Click 'create another app' at the bottom and select the 'script' radio button

Fill in the required details. For a personal script, the redirect URI is not important, set it to anything you like.

Take note of the app ID (a string with random characters) and the secret (the same but longer). These become, respectively, clientID and clientSecret.

username and password are the Reddit login credentials of the account you are using.

Note: Do NOT ever share your password or the client secret publicly!

See also: https://github.com/reddit-archive/reddit/wiki/OAuth2-Quick-Start-Example

Fields

ownerUsername :: Text

ownerPassword :: Text

ownerClientId :: Text

ownerClientSecret :: Text

CodeGrantCredentials
'Code grant' flow

Formally, this is the 'Authorisation Code Grant' for OAuth 2.0, which is described in Section 4.1 of IETF RFC 6749.

Begin by setting up a new Reddit app in the same way as for the 'resource owner' flow above. Then:

codeGrantClientId and codeGrantClientSecret are simply the 'app ID' and 'app secret' from the Reddit app.

codeGrantRedirectUri is the redirect URI you specified when setting up the app. Note that this must match exactly.

To obtain codeGrantCode, you must first construct the Reddit authentication link using mkRedditAuthURL. Send your user to this link, and they will be greeted with a Reddit page asking them to authorise your app. When they click 'allow', Reddit will send the user back to the redirect URI which you specified, along with a 'code' query parameter in the URI. You should then parse this and set it as codeGrantCode.

See also: https://github.com/reddit-archive/reddit/wiki/OAuth2

Fields

codeGrantClientId :: Text

codeGrantClientSecret :: Text

codeGrantRedirectUri :: Text

codeGrantCode :: Text

NoCredentials
Do not provide credentials. This is used when 'reconstructing' a RedditEnv from an existing token (e.g. one stored in a database) using mkEnvFromToken. You should not need to use this.

data Token

A parsed value of a Reddit access token.

Constructors

Token

Fields

token :: ByteString

tokenType :: ByteString

tokenExpiresAt :: UTCTime

tokenScopes :: Set Scope

tokenRefreshToken :: Maybe Text

Doesn't always exist

getToken :: Credentials -> ByteString -> IO Token

Get a token using credentials. This is the core of authenticate.

data AuthUrlParams

The information required for your app to construct a link for the user to 'authorise with Reddit'.

This is only required for the 'code grant' flow.

Constructors

AuthUrlParams

Fields

authUrlClientID :: Text

This is the client ID for your app.

authUrlState :: Text

This can be any text value you want, and must be different for every user you are authenticating. If the user authorises your app, they will be sent back to your redirect URI, and the same state will be included as a query parameter. You should check that the returned state matches the one which you specified here.

In principle, this is not mandatory. However, it is really bad practice to not use it, as it makes you vulnerable to CSRF attacks. See, e.g., Section 10.12 of the OAuth2 RFC. So, this library forces you to specify it.

authUrlRedirectUri :: Text

The URI which users will be sent back to after logging in. This must match the redirect URI set up in https://reddit.com/prefs/apps.

authUrlDuration :: Duration

Temporary or Permanent, depending on how long you need your user to be logged in for. Temporary tokens last for 1 hour, after which the user must re-authenticate if they want to continue using your app.

authUrlScopes :: Set Scope

The OAuth2 scopes that you need. This depends on what your app does on behalf of its users. See Scope for the full list.

mkRedditAuthURL

Arguments

:: Bool

-> AuthUrlParams
Whether to use the mobile version of the authorisation page

-> Text
The authorisation URL

Construct the Reddit authorisation URL that you should send your users to for the 'code grant' flow.

data Scope

Scopes are used to limit the things which a token can do. You should ensure that you only request a token for the scopes that you need. When authenticating with Reddit, the user will be able to see which scopes are being requested.

To see which scope(s) are required by each endpoint, see https://www.reddit.com/dev/api/oauth. Note that this library implements only a small subset of endpoints, so most of these are not very useful right now.

You can obtain this list of scopes via https://reddit.com/api/v1/scopes.

Constructors

ScopeCreddits
Spend the logged in user's Reddit gold.

ScopeModNote
Access mod notes for subreddits moderated by the logged in user.

ScopeModContributors
Add/remove users to/from approved user lists, ban/unban, and mute/unmute users on subreddits moderated by the logged in user.

ScopeModMail
Access and manage modmail for subreddits moderated by the logged in user.

ScopeModConfig
Manage the configuration, sidebar, and CSS for subreddits moderated by the logged in user.

ScopeSubscribe
Manage subreddit subscriptions as well as 'friends'.

ScopeStructuredStyles
Edit structured styles for subreddits moderated by the logged in user.

ScopeVote
Vote on posts and comments.

ScopeWikiEdit
Edit wiki pages.

ScopeMySubreddits
Access the list of subreddits moderated, contributed to, and subscribed to.

ScopeSubmit
Submit links and comments.

ScopeModLog
Access the moderation log in subreddits moderated by the logged in user.

ScopeModPosts
Approve, remove, mark as (N)SFW, and distinguish content in subreddits moderated by the logged in user.

ScopeModFlair
Manage and assign flair in subreddits moderated by the logged in user.

ScopeSave
Save and unsave posts and comments.

ScopeModOthers
Invite and remove moderators on subreddits moderated by the logged in user.

ScopeRead
Access posts and comments.

ScopePrivateMessages
Read and send private messages. (Note: this does not include chat. There is, at present, no way to access chat via the public Reddit API.)

ScopeReport
Report, hide, and unhide content.

ScopeIdentity
Get Reddit username and signup date.

ScopeLiveManage
Manage settings and contributors of live threads.

ScopeAccount
Update preferences and other account information. Does not grant access to email or password.

ScopeModTraffic
Access traffic stats in subreddits moderated by the user.

ScopeWikiRead
Read wiki pages.

ScopeEdit
Edit and delete posts and comments.

ScopeModWiki
Change editors and visibility of wiki pages in subreddits moderated by the logged in user.

ScopeModSelf
Accept invitations to moderate a subreddit and step down from moderating subreddits.

ScopeHistory
Access voting history, saved submissions, and hidden submissions.

ScopeFlair
Edit subreddit and post flair.

parseScopes :: Text -> Set Scope

Parse a canonical text list of scopes, which is part of the JSON returned by the Reddit authorisation server, into a set of known Scopes.

showScopes :: Set Scope -> Text

Construct a canonical text representation of a set of Scopes, which can be used as a query parameter for mkRedditAuthURL.

allScopes :: Set Scope

A quick way to get all scopes.

data Duration

This should be chosen depending on whether your app needs a temporary token (i.e. user must authenticate again after the token expires), or a permanent one (i.e. user does not need to authenticate again).

Internally, permanent tokens are handled by requesting a refresh token when the existing one expires.

Constructors

Temporary

Permanent